process

+++ src/routes/priceTypes.routes.js
@@ -4,6 +4,8 @@ const { saveAuditModel, decodedToken } = require("../utils");
 
 // Provider Model
 const PriceType = require("../models/priceType");
+//Product Model
+const Product = require("../models/product");
 
 // GET all PriceTypes
 router.get("/", async (req, res) => {
@@ -56,6 +58,21 @@ router.put("/:id", async (req, res) => {
 router.delete("/:id", async (req, res) => {
   try {
     const { _id } = decodedToken(req);
+
+    const products = await Product.find();
+
+    var ids = products.reduce(
+      (acc, obj) => [...acc, obj.prices[0].priceType._id],
+      []
+    );
+
+    if (ids.some((id) => id == req.params.id)) {
+      return res.status(500).send({
+        success: false,
+        message: "Error",
+        error: "Hay Productos con este precio",
+      });
+    }
 
     await PriceType.findByIdAndRemove(req.params.id);
